requirements.txt
annotated-types==0.7.0
anyio==4.8.0
base58==2.1.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
fastapi==0.115.8
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
solana==0.36.5
solders==0.25.0
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==14.2
